requirements.txt
# Core Streamlit and data processing
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0

# Visualization libraries
plotly>=5.15.0
altair>=5.0.0
matplotlib>=3.7.0

# Environment and API
python-dotenv>=1.0.0
openai>=1.0.0

# Snowflake integration (optional - uncomment if needed)
# streamlit[snowflake]>=1.28.0
# snowflake-snowpark-python>=1.8.0

# Additional utilities
requests>=2.31.0
